+++ libs/elasticsearch-nio/src/main/java/org/elasticsearch/nio/NioGroup.java
@@ -25,6 +25,7 @@ import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ThreadFactory;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -35,10 +36,9 @@ import java.util.stream.Stream;
 
 /**
  * The NioGroup is a group of selectors for interfacing with java nio. When it is started it will create the
- * configured number of socket and acceptor selectors. Each selector will be running in a dedicated thread.
- * Server connections can be bound using the {@link #bindServerChannel(InetSocketAddress, ChannelFactory)}
- * method. Client connections can be opened using the {@link #openChannel(InetSocketAddress, ChannelFactory)}
- * method.
+ * configured number of selectors. Each selector will be running in a dedicated thread. Server connections
+ * can be bound using the {@link #bindServerChannel(InetSocketAddress, ChannelFactory)} method. Client
+ * connections can be opened using the {@link #openChannel(InetSocketAddress, ChannelFactory)} method.
  * <p>
  * The logic specific to a particular channel is provided by the {@link ChannelFactory} passed to the method
  * when the channel is created. This is what allows an NioGroup to support different channel types.
@@ -46,35 +46,69 @@ import java.util.stream.Stream;
 public class NioGroup implements AutoCloseable {
 
 
-    private final ArrayList<NioSelector> acceptors;
+    private final List<NioSelector> dedicatedAcceptors;
     private final RoundRobinSupplier<NioSelector> acceptorSupplier;
 
-    private final ArrayList<NioSelector> socketSelectors;
-    private final RoundRobinSupplier<NioSelector> socketSelectorSupplier;
+    private final List<NioSelector> selectors;
+    private final RoundRobinSupplier<NioSelector> selectorSupplier;
 
     private final AtomicBoolean isOpen = new AtomicBoolean(true);
 
-    public NioGroup(ThreadFactory acceptorThreadFactory, int acceptorCount,
-                    Function<Supplier<NioSelector>, EventHandler> acceptorEventHandlerFunction,
-                    ThreadFactory socketSelectorThreadFactory, int socketSelectorCount,
-                    Supplier<EventHandler> socketEventHandlerFunction) throws IOException {
-        acceptors = new ArrayList<>(acceptorCount);
-        socketSelectors = new ArrayList<>(socketSelectorCount);
+    /**
+     * This will create an NioGroup with no dedicated acceptors. All server channels will be handled by the
+     * same selectors that are handling child channels.
+     *
+     * @param threadFactory factory to create selector threads
+     * @param selectorCount the number of selectors to be created
+     * @param eventHandlerFunction function for creating event handlers
+     * @throws IOException occurs if there is a problem while opening a java.nio.Selector
+     */
+    public NioGroup(ThreadFactory threadFactory, int selectorCount, Function<Supplier<NioSelector>, EventHandler> eventHandlerFunction)
+        throws IOException {
+        this(null, 0, threadFactory, selectorCount, eventHandlerFunction);
+
+
+    }
+
+    /**
+     * This will create an NioGroup with dedicated acceptors. All server channels will be handled by a group
+     * of selectors dedicated to accepting channels. These accepted channels will be handed off the
+     * non-server selectors.
+     *
+     * @param acceptorThreadFactory factory to create acceptor selector threads
+     * @param dedicatedAcceptorCount the number of dedicated acceptor selectors to be created
+     * @param selectorThreadFactory factory to create non-acceptor selector threads
+     * @param selectorCount the number of non-acceptor selectors to be created
+     * @param eventHandlerFunction function for creating event handlers
+     * @throws IOException occurs if there is a problem while opening a java.nio.Selector
+     */
+    public NioGroup(ThreadFactory acceptorThreadFactory, int dedicatedAcceptorCount, ThreadFactory selectorThreadFactory, int selectorCount,
+                    Function<Supplier<NioSelector>, EventHandler> eventHandlerFunction) throws IOException {
+        dedicatedAcceptors = new CopyOnWriteArrayList<>();
+        selectors = new CopyOnWriteArrayList<>();
 
         try {
-            for (int i = 0; i < socketSelectorCount; ++i) {
-                NioSelector selector = new NioSelector(socketEventHandlerFunction.get());
-                socketSelectors.add(selector);
+            for (int i = 0; i < selectorCount; ++i) {
+                NioSelector selector = new NioSelector(eventHandlerFunction.apply(new RoundRobinSupplier<>(selectors, selectorCount)));
+                selectors.add(selector);
             }
-            startSelectors(socketSelectors, socketSelectorThreadFactory);
 
-            for (int i = 0; i < acceptorCount; ++i) {
-                NioSelector[] childSelectors = this.socketSelectors.toArray(new NioSelector[this.socketSelectors.size()]);
-                Supplier<NioSelector> selectorSupplier = new RoundRobinSupplier<>(childSelectors);
-                NioSelector acceptor = new NioSelector(acceptorEventHandlerFunction.apply(selectorSupplier));
-                acceptors.add(acceptor);
+            for (int i = 0; i < dedicatedAcceptorCount; ++i) {
+                NioSelector acceptor = new NioSelector(eventHandlerFunction.apply(new RoundRobinSupplier<>(selectors, selectorCount)));
+                dedicatedAcceptors.add(acceptor);
             }
-            startSelectors(acceptors, acceptorThreadFactory);
+
+            startSelectors(selectors, selectorThreadFactory);
+            startSelectors(dedicatedAcceptors, acceptorThreadFactory);
+
+            if (dedicatedAcceptorCount != 0) {
+                acceptorSupplier = new RoundRobinSupplier<>(dedicatedAcceptors, dedicatedAcceptorCount);
+            } else {
+                acceptorSupplier = new RoundRobinSupplier<>(selectors, selectorCount);
+            }
+            selectorSupplier = new RoundRobinSupplier<>(selectors, selectorCount);
+            assert selectorCount == selectors.size() : "We need to have created all the selectors at this point.";
+            assert dedicatedAcceptorCount == dedicatedAcceptors.size() : "We need to have created all the acceptors at this point.";
         } catch (Exception e) {
             try {
                 close();
@@ -83,29 +117,23 @@ public class NioGroup implements AutoCloseable {
             }
             throw e;
         }
-
-        socketSelectorSupplier = new RoundRobinSupplier<>(socketSelectors.toArray(new NioSelector[socketSelectors.size()]));
-        acceptorSupplier = new RoundRobinSupplier<>(acceptors.toArray(new NioSelector[acceptors.size()]));
     }
 
     public <S extends NioServerSocketChannel> S bindServerChannel(InetSocketAddress address, ChannelFactory<S, ?> factory)
         throws IOException {
         ensureOpen();
-        if (acceptors.isEmpty()) {
-            throw new IllegalArgumentException("There are no acceptors configured. Without acceptors, server channels are not supported.");
-        }
         return factory.openNioServerSocketChannel(address, acceptorSupplier);
     }
 
     public <S extends NioSocketChannel> S openChannel(InetSocketAddress address, ChannelFactory<?, S> factory) throws IOException {
         ensureOpen();
-        return factory.openNioChannel(address, socketSelectorSupplier);
+        return factory.openNioChannel(address, selectorSupplier);
     }
 
     @Override
     public void close() throws IOException {
         if (isOpen.compareAndSet(true, false)) {
-            List<NioSelector> toClose = Stream.concat(acceptors.stream(), socketSelectors.stream()).collect(Collectors.toList());
+            List<NioSelector> toClose = Stream.concat(dedicatedAcceptors.stream(), selectors.stream()).collect(Collectors.toList());
             List<IOException> closingExceptions = new ArrayList<>();
             for (NioSelector selector : toClose) {
                 try {
