+++ libs/elasticsearch-nio/src/test/java/org/elasticsearch/nio/NioGroupTests.java
@@ -38,9 +38,8 @@ public class NioGroupTests extends ESTestCase {
     @SuppressWarnings("unchecked")
     public void setUp() throws Exception {
         super.setUp();
-        nioGroup = new NioGroup(daemonThreadFactory(Settings.EMPTY, "acceptor"), 1,
-            (s) -> new EventHandler(mock(Consumer.class), s), daemonThreadFactory(Settings.EMPTY, "selector"), 1,
-            () -> new EventHandler(mock(Consumer.class)));
+        nioGroup = new NioGroup(daemonThreadFactory(Settings.EMPTY, "acceptor"), 1, daemonThreadFactory(Settings.EMPTY, "selector"), 1,
+            (s) -> new EventHandler(mock(Consumer.class), s));
     }
 
     @Override
@@ -76,8 +75,8 @@ public class NioGroupTests extends ESTestCase {
     public void testExceptionAtStartIsHandled() throws IOException {
         RuntimeException ex = new RuntimeException();
         CheckedRunnable<IOException> ctor = () -> new NioGroup(r -> {throw ex;}, 1,
-            (s) -> new EventHandler(mock(Consumer.class), s), daemonThreadFactory(Settings.EMPTY, "selector"),
-            1, () -> new EventHandler(mock(Consumer.class)));
+            daemonThreadFactory(Settings.EMPTY, "selector"),
+            1, (s) -> new EventHandler(mock(Consumer.class), s));
         RuntimeException runtimeException = expectThrows(RuntimeException.class, ctor::run);
         assertSame(ex, runtimeException);
         // ctor starts threads. So we are testing that a failure to construct will stop threads. Our thread
