+++ server/src/main/java/org/elasticsearch/transport/Transports.java
@@ -30,7 +30,6 @@ public enum Transports {
     public static final String TEST_MOCK_TRANSPORT_THREAD_PREFIX = "__mock_network_thread";
 
     public static final String NIO_TRANSPORT_WORKER_THREAD_NAME_PREFIX = "es_nio_transport_worker";
-    public static final String NIO_TRANSPORT_ACCEPTOR_THREAD_NAME_PREFIX = "es_nio_transport_acceptor";
 
     /**
      * Utility method to detect whether a thread is a network thread. Typically
@@ -44,8 +43,7 @@ public enum Transports {
                 TcpTransport.TRANSPORT_SERVER_WORKER_THREAD_NAME_PREFIX,
                 TcpTransport.TRANSPORT_CLIENT_BOSS_THREAD_NAME_PREFIX,
                 TEST_MOCK_TRANSPORT_THREAD_PREFIX,
-                NIO_TRANSPORT_WORKER_THREAD_NAME_PREFIX,
-                NIO_TRANSPORT_ACCEPTOR_THREAD_NAME_PREFIX)) {
+                NIO_TRANSPORT_WORKER_THREAD_NAME_PREFIX)) {
             if (threadName.contains(s)) {
                 return true;
             }
