+++ test/framework/src/main/java/org/elasticsearch/transport/nio/TestingSocketEventHandler.java
@@ -20,6 +20,7 @@
 package org.elasticsearch.transport.nio;
 
 import org.elasticsearch.nio.EventHandler;
+import org.elasticsearch.nio.NioSelector;
 import org.elasticsearch.nio.SocketChannelContext;
 
 import java.io.IOException;
@@ -27,13 +28,14 @@ import java.util.Collections;
 import java.util.Set;
 import java.util.WeakHashMap;
 import java.util.function.Consumer;
+import java.util.function.Supplier;
 
 public class TestingSocketEventHandler extends EventHandler {
 
     private Set<SocketChannelContext> hasConnectedMap = Collections.newSetFromMap(new WeakHashMap<>());
 
-    public TestingSocketEventHandler(Consumer<Exception> exceptionHandler) {
-        super(exceptionHandler);
+    public TestingSocketEventHandler(Consumer<Exception> exceptionHandler, Supplier<NioSelector> selectorSupplier) {
+        super(exceptionHandler, selectorSupplier);
     }
 
     public void handleConnect(SocketChannelContext context) throws IOException {
